Split template dep mgmt into two phases
New problem is when we are testing the libraries for validity,
sometimes we want to give an explicit argument to the 'mapsto'
optional arg w/n a dependency. However, we currently can't give
a binding to a nested dependency. The result is that we end up with
multiple mapsto (because the 'optional' mapsto argument gets a newly
minted URI from template.fill()), which fails validation.

This commit splits template dependency management into 2 phases: an
initial phase which adds the dependency links and args to the DB; a
second phase which checks that the args/params between the template and
its dependencies are all valid and work as expected.