Add new transitive_parameters method to accelerate library loading
The prior implementation of check_template_dependency_relationship used
inline_dependencies to check the parameters of template dependencies.
inline_dependencies can be slow-ish because it copies the bodies of
templates (these are graphs, so they inherit the performance
characteristics of the underlying store). We can compute the parameters
through the dependency tree automatically without copying the graphs,
which is much faster.